Feature 12 : Create Base Class (models) for SQLAlchemy (#39)
* refactor(back-end): create new files and refactor project before creating models

* refactor(back-end): move file content in new created files

* refactor(back-end): delete useless files

* refactor(back-end): move params to configs file

* feat(database): add SQLAlchemy models to backend

* feat(database): add Mapped type

* feat(back-end): add class models

* refactor(back-end): remove useless import
