+++ back-end/src/main/app.py
@@ -1,31 +1,26 @@
 from flask import Flask
-from flask_sqlalchemy import SQLAlchemy
-from sqlalchemy.orm import DeclarativeBase
 
-from app_qrcode import app_qrcode
-from app_guest_book import app_guest_book
-
-
-# db config
-class Base(DeclarativeBase):
-    pass
-
-db = SQLAlchemy(model_class=Base)
+from endpoints import guest_book_blueprint, qrcode_blueprint
+from models import db
+from configs import Database, API
 
 
-
+# create Flask app
 app = Flask(__name__)
-app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///project.db"
+app.config["SQLALCHEMY_DATABASE_URI"] = Database.URI
+
 # initialize the app with the extension
 db.init_app(app)
-# register blueprint
-app.register_blueprint(app_guest_book)
-app.register_blueprint(app_qrcode)
 
+# create tables if not created yet
+with app.app_context():
+    db.create_all()
 
-
+# register blueprints
+app.register_blueprint(guest_book_blueprint)
+app.register_blueprint(qrcode_blueprint)
 
 
 # launch
 if __name__ == '__main__':
-    app.run(host='0.0.0.0', port=5000)
+    app.run(host=API.HOST, port=API.PORT)
