+++ back-end/src/main/configs.py
@@ -1,0 +1,7 @@
+class Database:
+    URI = "sqlite:///project.db"
+
+
+class API:
+    HOST = "0.0.0.0"
+    PORT = 5000
